adds ADR, addresses GH lint
