+++ src/pkg/bundler/bundler.go
@@ -5,74 +5,46 @@
 package bundler
 
 import (
-	"errors"
-	"fmt"
-	"strings"
-
 	"github.com/defenseunicorns/uds-cli/src/types"
-	"github.com/defenseunicorns/zarf/src/pkg/message"
-	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
-	"oras.land/oras-go/v2/registry"
 )
 
 type Bundler struct {
-	bundle     *types.UDSBundle
-	createOpts types.BundleCreateOptions
-	tmpDstDir  string
+	bundle    *types.UDSBundle
+	output    string
+	tmpDstDir string
 }
 
 type Pusher interface {
 }
 
 type Options struct {
-	// todo comment all these options
-	Bundle     *types.UDSBundle
-	CreateOpts types.BundleCreateOptions
-	TmpDstDir  string
+	Bundle    *types.UDSBundle
+	Output    string
+	TmpDstDir string
 }
 
 func NewBundler(opts *Options) *Bundler {
 	b := Bundler{
-		bundle:     opts.Bundle,
-		createOpts: opts.CreateOpts,
-		tmpDstDir:  opts.TmpDstDir,
+		bundle:    opts.Bundle,
+		output:    opts.Output,
+		tmpDstDir: opts.TmpDstDir,
 	}
 	return &b
 }
 
 func (b *Bundler) Create() error {
-	if b.createOpts.Output == "" {
-		err := b.createLocalBundle(nil)
+	if b.output == "" {
+		localBundle := NewLocalBundle(&LocalBundleOpts{Bundle: b.bundle, TmpDstDir: b.tmpDstDir})
+		err := localBundle.create(nil)
 		if err != nil {
 			return err
 		}
 	} else {
-		err := b.createRemoteBundle(nil)
+		remoteBundle := NewRemoteBundle(&RemoteBundleOpts{Bundle: b.bundle, Output: b.output})
+		err := remoteBundle.create(nil)
 		if err != nil {
 			return err
 		}
 	}
 	return nil
-}
-
-// copied from: https://github.com/defenseunicorns/zarf/blob/main/src/pkg/oci/utils.go
-func referenceFromMetadata(registryLocation string, metadata *types.UDSMetadata) (string, error) {
-	ver := metadata.Version
-	if len(ver) == 0 {
-		return "", errors.New("version is required for publishing")
-	}
-
-	if !strings.HasSuffix(registryLocation, "/") {
-		registryLocation = registryLocation + "/"
-	}
-	registryLocation = strings.TrimPrefix(registryLocation, helpers.OCIURLPrefix)
-	raw := fmt.Sprintf("%s%s:%s", registryLocation, metadata.Name, ver)
-
-	message.Debug("Raw OCI reference from metadata:", raw)
-	ref, err := registry.ParseReference(raw)
-	if err != nil {
-		return "", err
-	}
-
-	return ref.String(), nil
 }
