+++ src/pkg/bundler/common.go
@@ -1,0 +1,84 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2023-Present The UDS Authors
+
+// Package bundler defines behavior for bundling packages
+package bundler
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+
+	"github.com/defenseunicorns/uds-cli/src/pkg/utils"
+	"github.com/defenseunicorns/uds-cli/src/types"
+	"github.com/defenseunicorns/zarf/src/pkg/message"
+	"github.com/defenseunicorns/zarf/src/pkg/oci"
+	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	"oras.land/oras-go/v2/registry"
+)
+
+// copied from: https://github.com/defenseunicorns/zarf/blob/main/src/pkg/oci/push.go
+func manifestAnnotationsFromMetadata(metadata *types.UDSMetadata) map[string]string {
+	annotations := map[string]string{
+		ocispec.AnnotationDescription: metadata.Description,
+	}
+
+	if url := metadata.URL; url != "" {
+		annotations[ocispec.AnnotationURL] = url
+	}
+	if authors := metadata.Authors; authors != "" {
+		annotations[ocispec.AnnotationAuthors] = authors
+	}
+	if documentation := metadata.Documentation; documentation != "" {
+		annotations[ocispec.AnnotationDocumentation] = documentation
+	}
+	if source := metadata.Source; source != "" {
+		annotations[ocispec.AnnotationSource] = source
+	}
+	if vendor := metadata.Vendor; vendor != "" {
+		annotations[ocispec.AnnotationVendor] = vendor
+	}
+
+	return annotations
+}
+
+// copied from: https://github.com/defenseunicorns/zarf/blob/main/src/pkg/oci/push.go
+func pushManifestConfigFromMetadata(r *oci.OrasRemote, metadata *types.UDSMetadata, build *types.UDSBuildData) (ocispec.Descriptor, error) {
+	annotations := map[string]string{
+		ocispec.AnnotationTitle:       metadata.Name,
+		ocispec.AnnotationDescription: metadata.Description,
+	}
+	manifestConfig := oci.ConfigPartial{
+		Architecture: build.Architecture,
+		OCIVersion:   "1.0.1",
+		Annotations:  annotations,
+	}
+	manifestConfigDesc, err := utils.ToOCIRemote(manifestConfig, oci.ZarfLayerMediaTypeBlob, r)
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+	return manifestConfigDesc, nil
+}
+
+// copied from: https://github.com/defenseunicorns/zarf/blob/main/src/pkg/oci/utils.go
+func referenceFromMetadata(registryLocation string, metadata *types.UDSMetadata) (string, error) {
+	ver := metadata.Version
+	if len(ver) == 0 {
+		return "", errors.New("version is required for publishing")
+	}
+
+	if !strings.HasSuffix(registryLocation, "/") {
+		registryLocation = registryLocation + "/"
+	}
+	registryLocation = strings.TrimPrefix(registryLocation, helpers.OCIURLPrefix)
+	raw := fmt.Sprintf("%s%s:%s", registryLocation, metadata.Name, ver)
+
+	message.Debug("Raw OCI reference from metadata:", raw)
+	ref, err := registry.ParseReference(raw)
+	if err != nil {
+		return "", err
+	}
+
+	return ref.String(), nil
+}
