+++ src/pkg/bundler/localbundle.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/defenseunicorns/uds-cli/src/config"
 	"github.com/defenseunicorns/uds-cli/src/pkg/bundler/fetcher"
-	"github.com/defenseunicorns/uds-cli/src/pkg/bundler/pusher"
 	"github.com/defenseunicorns/uds-cli/src/pkg/utils"
 	"github.com/defenseunicorns/uds-cli/src/types"
 	"github.com/defenseunicorns/zarf/src/pkg/message"
@@ -27,13 +26,30 @@ import (
 	ocistore "oras.land/oras-go/v2/content/oci"
 )
 
-// createLocalBundle creates the bundle and outputs to a local tarball
-func (b *Bundler) createLocalBundle(signature []byte) error {
-	bundle := b.bundle
+type LocalBundleOpts struct {
+	Bundle    *types.UDSBundle
+	TmpDstDir string
+}
+
+type LocalBundle struct {
+	bundle    *types.UDSBundle
+	tmpDstDir string
+}
+
+func NewLocalBundle(opts *LocalBundleOpts) *LocalBundle {
+	return &LocalBundle{
+		bundle:    opts.Bundle,
+		tmpDstDir: opts.TmpDstDir,
+	}
+}
+
+// create creates the bundle and outputs to a local tarball
+func (lo *LocalBundle) create(signature []byte) error {
+	bundle := lo.bundle
 	if bundle.Metadata.Architecture == "" {
 		return fmt.Errorf("architecture is required for bundling")
 	}
-	store, err := ocistore.NewWithContext(context.TODO(), b.tmpDstDir)
+	store, err := ocistore.NewWithContext(context.TODO(), lo.tmpDstDir)
 	ctx := context.TODO()
 
 	message.HeaderInfof("🐕 Fetching Packages")
@@ -46,12 +62,12 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 	fetcherConfig := fetcher.Config{
 		Bundle:             bundle,
 		Store:              store,
-		TmpDstDir:          b.tmpDstDir,
-		NumPkgs:            len(b.bundle.Packages),
+		TmpDstDir:          lo.tmpDstDir,
+		NumPkgs:            len(lo.bundle.Packages),
 		BundleRootManifest: &rootManifest,
 	}
 
-	message.Debug("Bundling", bundle.Metadata.Name, "to", b.tmpDstDir)
+	message.Debug("Bundling", bundle.Metadata.Name, "to", lo.tmpDstDir)
 	if err != nil {
 		return err
 	}
@@ -73,7 +89,7 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 		// todo: if we know the path to where the blobs are stored, we can use that instead of the artifactPathMap?
 		for _, layer := range layerDescs {
 			digest := layer.Digest.Encoded()
-			artifactPathMap[filepath.Join(b.tmpDstDir, config.BlobsDir, digest)] = filepath.Join(config.BlobsDir, digest)
+			artifactPathMap[filepath.Join(lo.tmpDstDir, config.BlobsDir, digest)] = filepath.Join(config.BlobsDir, digest)
 		}
 	}
 
@@ -88,7 +104,7 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 	// append uds-bundle.yaml layer to rootManifest and grab path for archiving
 	rootManifest.Layers = append(rootManifest.Layers, bundleYAMLDesc)
 	digest := bundleYAMLDesc.Digest.Encoded()
-	artifactPathMap[filepath.Join(b.tmpDstDir, config.BlobsDir, digest)] = filepath.Join(config.BlobsDir, digest)
+	artifactPathMap[filepath.Join(lo.tmpDstDir, config.BlobsDir, digest)] = filepath.Join(config.BlobsDir, digest)
 
 	// create and push bundle manifest config
 	manifestConfigDesc, err := pushManifestConfig(store, bundle.Metadata, bundle.Build)
@@ -96,7 +112,7 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 		return err
 	}
 	manifestConfigDigest := manifestConfigDesc.Digest.Encoded()
-	artifactPathMap[filepath.Join(b.tmpDstDir, config.BlobsDir, manifestConfigDigest)] = filepath.Join(config.BlobsDir, manifestConfigDigest)
+	artifactPathMap[filepath.Join(lo.tmpDstDir, config.BlobsDir, manifestConfigDigest)] = filepath.Join(config.BlobsDir, manifestConfigDigest)
 
 	rootManifest.Config = manifestConfigDesc
 	rootManifest.SchemaVersion = 2
@@ -106,13 +122,13 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 		return err
 	}
 	digest = rootManifestDesc.Digest.Encoded()
-	artifactPathMap[filepath.Join(b.tmpDstDir, config.BlobsDir, digest)] = filepath.Join(config.BlobsDir, digest)
+	artifactPathMap[filepath.Join(lo.tmpDstDir, config.BlobsDir, digest)] = filepath.Join(config.BlobsDir, digest)
 
 	// grab index.json
-	artifactPathMap[filepath.Join(b.tmpDstDir, "index.json")] = "index.json"
+	artifactPathMap[filepath.Join(lo.tmpDstDir, "index.json")] = "index.json"
 
 	// grab oci-layout
-	artifactPathMap[filepath.Join(b.tmpDstDir, "oci-layout")] = "oci-layout"
+	artifactPathMap[filepath.Join(lo.tmpDstDir, "oci-layout")] = "oci-layout"
 
 	// push the bundle's signature todo: need to understand functionality and add tests
 	if len(signature) > 0 {
@@ -131,7 +147,7 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 		return err
 	}
 	// ensure the bundle root manifest is the only manifest in the index.json
-	err = cleanIndexJSON(b.tmpDstDir, rootManifestDesc)
+	err = cleanIndexJSON(lo.tmpDstDir, rootManifestDesc)
 	if err != nil {
 		return err
 	}
@@ -143,158 +159,6 @@ func (b *Bundler) createLocalBundle(signature []byte) error {
 	}
 
 	return nil
-}
-
-// createRemoteBundle creates the bundle in a remote OCI registry publishes w/ optional signature to the remote repository.
-func (b *Bundler) createRemoteBundle(signature []byte) error {
-	// set the bundle remote's reference from metadata
-	b.createOpts.Output = utils.EnsureOCIPrefix(b.createOpts.Output)
-	ref, err := referenceFromMetadata(b.createOpts.Output, &b.bundle.Metadata)
-	if err != nil {
-		return err
-	}
-	platform := ocispec.Platform{
-		Architecture: config.GetArch(),
-		OS:           oci.MultiOS,
-	}
-
-	// create the bundle remote
-	bundleRemote, err := oci.NewOrasRemote(ref, platform)
-	if err != nil {
-		return err
-	}
-	bundle := b.bundle
-	if bundle.Metadata.Architecture == "" {
-		return fmt.Errorf("architecture is required for bundling")
-	}
-	dstRef := bundleRemote.Repo().Reference
-	message.Debug("Bundling", bundle.Metadata.Name, "to", dstRef)
-
-	rootManifest := ocispec.Manifest{}
-	pusherConfig := pusher.Config{
-		Bundle:    bundle,
-		RemoteDst: bundleRemote,
-		NumPkgs:   len(bundle.Packages),
-	}
-
-	for i, pkg := range bundle.Packages {
-		// todo: can leave this block here or move to pusher.NewPkgPusher (would be closer to NewPkgFetcher pattern)
-		pkgUrl := fmt.Sprintf("%s:%s", pkg.Repository, pkg.Ref)
-		src, err := oci.NewOrasRemote(pkgUrl, platform)
-		if err != nil {
-			return err
-		}
-		pusherConfig.RemoteSrc = src
-		pkgRootManifest, err := src.FetchRoot()
-		if err != nil {
-			return err
-		}
-		pusherConfig.PkgRootManifest = pkgRootManifest
-		pusherConfig.PkgIter = i
-
-		remotePusher := pusher.NewPkgPusher(pkg, pusherConfig)
-		zarfManifestDesc, err := remotePusher.Push()
-		if err != nil {
-			return err
-		}
-		rootManifest.Layers = append(rootManifest.Layers, zarfManifestDesc)
-	}
-
-	// push the bundle's metadata
-	bundleYamlBytes, err := goyaml.Marshal(bundle)
-	if err != nil {
-		return err
-	}
-	bundleYamlDesc, err := bundleRemote.PushLayer(bundleYamlBytes, oci.ZarfLayerMediaTypeBlob)
-	if err != nil {
-		return err
-	}
-	bundleYamlDesc.Annotations = map[string]string{
-		ocispec.AnnotationTitle: config.BundleYAML,
-	}
-
-	message.Debug("Pushed", config.BundleYAML+":", message.JSONValue(bundleYamlDesc))
-	rootManifest.Layers = append(rootManifest.Layers, bundleYamlDesc)
-
-	// push the bundle's signature
-	if len(signature) > 0 {
-		bundleYamlSigDesc, err := bundleRemote.PushLayer(signature, oci.ZarfLayerMediaTypeBlob)
-		if err != nil {
-			return err
-		}
-		bundleYamlSigDesc.Annotations = map[string]string{
-			ocispec.AnnotationTitle: config.BundleYAMLSignature,
-		}
-		rootManifest.Layers = append(rootManifest.Layers, bundleYamlSigDesc)
-		message.Debug("Pushed", config.BundleYAMLSignature+":", message.JSONValue(bundleYamlSigDesc))
-	}
-
-	// push the bundle manifest config
-	configDesc, err := pushManifestConfigFromMetadata(bundleRemote, &bundle.Metadata, &bundle.Build)
-	if err != nil {
-		return err
-	}
-
-	message.Debug("Pushed config:", message.JSONValue(configDesc))
-
-	// check for existing index
-	index, err := utils.GetIndex(bundleRemote, dstRef.String())
-	if err != nil {
-		return err
-	}
-
-	// push bundle root manifest
-	rootManifest.Config = configDesc
-	rootManifest.SchemaVersion = 2
-	rootManifest.Annotations = manifestAnnotationsFromMetadata(&bundle.Metadata) // maps to registry UI
-	rootManifestDesc, err := utils.ToOCIRemote(rootManifest, ocispec.MediaTypeImageManifest, bundleRemote)
-	if err != nil {
-		return err
-	}
-
-	// create or update, then push index.json
-	err = utils.UpdateIndex(index, bundleRemote, bundle, rootManifestDesc)
-	if err != nil {
-		return err
-	}
-
-	message.HorizontalRule()
-	flags := ""
-	if config.CommonOptions.Insecure {
-		flags = "--insecure"
-	}
-	message.Title("To inspect/deploy/pull:", "")
-	message.Command("inspect oci://%s %s", dstRef, flags)
-	message.Command("deploy oci://%s %s", dstRef, flags)
-	message.Command("pull oci://%s %s", dstRef, flags)
-
-	return nil
-}
-
-// duplicated in bundle.go!
-// copied from: https://github.com/defenseunicorns/zarf/blob/main/src/pkg/oci/push.go
-func manifestAnnotationsFromMetadata(metadata *types.UDSMetadata) map[string]string {
-	annotations := map[string]string{
-		ocispec.AnnotationDescription: metadata.Description,
-	}
-
-	if url := metadata.URL; url != "" {
-		annotations[ocispec.AnnotationURL] = url
-	}
-	if authors := metadata.Authors; authors != "" {
-		annotations[ocispec.AnnotationAuthors] = authors
-	}
-	if documentation := metadata.Documentation; documentation != "" {
-		annotations[ocispec.AnnotationDocumentation] = documentation
-	}
-	if source := metadata.Source; source != "" {
-		annotations[ocispec.AnnotationSource] = source
-	}
-	if vendor := metadata.Vendor; vendor != "" {
-		annotations[ocispec.AnnotationVendor] = vendor
-	}
-
-	return annotations
 }
 
 // pushBundleYAMLToStore pushes the uds-bundle.yaml to a provided OCI store
@@ -440,22 +304,4 @@ func cleanIndexJSON(tmpDir string, bundleRootDesc ocispec.Descriptor) error {
 		return err
 	}
 	return nil
-}
-
-// copied from: https://github.com/defenseunicorns/zarf/blob/main/src/pkg/oci/push.go
-func pushManifestConfigFromMetadata(r *oci.OrasRemote, metadata *types.UDSMetadata, build *types.UDSBuildData) (ocispec.Descriptor, error) {
-	annotations := map[string]string{
-		ocispec.AnnotationTitle:       metadata.Name,
-		ocispec.AnnotationDescription: metadata.Description,
-	}
-	manifestConfig := oci.ConfigPartial{
-		Architecture: build.Architecture,
-		OCIVersion:   "1.0.1",
-		Annotations:  annotations,
-	}
-	manifestConfigDesc, err := utils.ToOCIRemote(manifestConfig, oci.ZarfLayerMediaTypeBlob, r)
-	if err != nil {
-		return ocispec.Descriptor{}, err
-	}
-	return manifestConfigDesc, nil
 }
